Create new monitor route for clients
