+++ ui/app/components/client-subnav.js
@@ -1,0 +1,5 @@
+import Component from '@ember/component';
+import { tagName } from '@ember-decorators/component';
+
+@tagName('')
+export default class ClientSubnav extends Component {}
