+++ ui/app/controllers/clients/client/index.js
