+++ ui/app/templates/allocations/allocation/index.hbs
@@ -1,5 +1,5 @@
 {{title "Allocation " model.name}}
-{{allocation-subnav allocation=model}}
+<AllocationSubnav @allocation={{model}} />
 <section class="section">
   {{#if error}}
     <div data-test-inline-error class="notification is-danger">
