+++ ui/app/templates/clients/client/monitor.hbs
@@ -1,0 +1,4 @@
+{{title "Client " (or model.name model.shortId)}}
+<ClientSubnav @client={{model}} />
+<section class="section">
+</section>
