+++ ui/app/templates/components/client-subnav.hbs
@@ -1,0 +1,6 @@
+<div class="tabs is-subnav">
+  <ul>
+    <li>{{#link-to "clients.client.index" client activeClass="is-active"}}Overview{{/link-to}}</li>
+    <li>{{#link-to "clients.client.monitor" client activeClass="is-active"}}Monitor{{/link-to}}</li>
+  </ul>
+</div>
